linkTools file divided into two files for android and ios
